Move StokesSpectralCube to dedicated file
