+++ spectral_cube/__init__.py
@@ -10,5 +10,6 @@ This is an Astropy affiliated package.
 from ._astropy_init import *
 # ----------------------------------------------------------------------------
 if not _ASTROPY_SETUP_:
-    from .spectral_cube import SpectralCube, StokesSpectralCube, Projection
+    from .spectral_cube import SpectralCube
+    from .stokes_spectral_cube import StokesSpectralCube
     from .masks import *
